+++ resources/views/roles/index.blade.php
@@ -2,13 +2,16 @@
 <div class="card card-contact">
     <div class="pl-3 row align-items-center">
         <div class="card-body">
-            <div class="card-title mr-3 col-xs">
-                @if($role->created_by != 1 && Gate::check('edit permission'))
+            <div class="card-title d-flex align-items-center mr-3 col-xs">
+                @if(Gate::check('edit permission'))
                     <a href="{{ route('roles.edit',$role->id) }}" data-remote="true" data-type="text">
                         <h6 data-filter-by="text">{{ $role->name }}</h6>
                     </a>
                 @else
                     <h6 data-filter-by="text">{{ $role->name }}</h6>
+                @endif
+                @if($role->created_by == 1)
+                    <span class="ml-3 badge badge-secondary">default</span>
                 @endif
             </div>
             <div class="card-meta col-xl">
@@ -18,7 +21,6 @@
                     @endfor
                 </div>
             </div>
-            @if($role->created_by != 1)
             <div class="dropdown card-options">
                 <button class="btn-options" type="button" id="task-dropdown-button-1" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
                     <i class="material-icons">more_vert</i>
@@ -30,15 +32,14 @@
                         <span>{{__('Edit')}}</span>
                     </a>
                     @endcan
-                    @can('delete permission')
+                    @if($role->created_by != 1 && Gate::check('delete permission'))
                         <div class="dropdown-divider"></div>
                         <a class="dropdown-item text-danger" href="{{ route('roles.destroy', $role->id) }}" data-method="delete" data-remote="true" data-type="text">
                             <span>{{__('Delete')}}</span>
                         </a>
-                    @endcan
+                    @endif
                 </div>
             </div>
-            @endif
         </div>
     </div>
 </div>
